refactor: move exchange rate APY to helper, remove dead code
- Move computeExchangeRateAPY to earnYield.helper.ts (follows existing pattern)
- Remove dead computeYieldAPYFromAnchors, computeRatioFromRaw, getHistoryValueAtAnchor
- Remove unused baseAddress from yieldBenchmarks config
